add travis
update travis etc.

install root

update travis

install root

update

alias for readlink in osx

further ls

put before install in correct position

ls..

give up -release folder

before_install is in osx..

add coverage test

all plain

set signature

use .

modern openmp

update also Config.cmake

avoid using readlink -f

folder does not exist

fix a few mistakes

add coverage test. bin_lo can be negative

use libomp

add badge

+++ PDFs/src/HistogramPdf.cu
@@ -32,7 +32,7 @@ EXEC_TARGET fptype device_ScaleShiftEvalRawHistogram (fptype* evt, fptype* p, un
   int bin_lo = (int) FLOOR(bin_fp);
   const int wid = RO_CACHE(indices[1]);
   if(bin_lo<0) return RO_CACHE((dev_raw_histograms[wid])[0]);
-  if(bin_lo>=RO_CACHE(indices[5])/*bin_max*/) return RO_CACHE((dev_raw_histograms[wid])[bin_lo]);
+  if(static_cast<unsigned int>(bin_lo)>=RO_CACHE(indices[5])/*bin_max*/) return RO_CACHE((dev_raw_histograms[wid])[bin_lo]);
   const fptype y_lo = RO_CACHE((dev_raw_histograms[wid])[bin_lo]);
   const fptype y_up = RO_CACHE((dev_raw_histograms[wid])[bin_lo+1]);
   return y_lo+(y_up-y_lo)*(bin_fp-bin_lo);
